B.W.M. , Bug fixes, Destroy Matches
We can now create a board without matchnes.
We fixed memory leak bugs on swipe right and up.
We fixed a bug that could put us in an infinite loop on Node Creation.
We Destroy now destroy our matches and leave a null space where we had a node/Dot